added authorization

+++ Controllers/hall.controller.js
@@ -1,6 +1,8 @@
 // Helps the admin to save the new halls in the database
-const HallModel = require("../Models/hall.model");
-const { isSuperAdmin } = require("../Controllers/auth.controller");
+const HallModel = require('../Models/hall.model')
+const { isSuperAdmin } = require('../Controllers/auth.controller')
+// const { verify } = require('jsonwebtoken')
+const { verify1 } = require('./user.controller')
 async function addHall(hall1, req, res) {
   if (isSuperAdmin()) {
     const {
@@ -16,13 +18,13 @@ async function addHall(hall1, req, res) {
       hall_catering,
       hall_duration,
       hall_rating,
-    } = hall1;
+    } = hall1
 
     //if the hall id is repeated reject the request
-    const hall = await HallModel.findOne({ hall_id: hall_id });
+    const hall = await HallModel.findOne({ hall_id: hall_id })
     if (hall) {
-      res.status(404).send("Hall id is taken!");
-      return;
+      res.status(404).send('Hall id is taken!')
+      return
     }
 
     if (
@@ -51,56 +53,61 @@ async function addHall(hall1, req, res) {
         hall_catering,
         hall_duration,
         hall_rating,
-      });
-      res.status(200).json(hall);
-      return;
+      })
+      res.status(200).json(hall)
+      return
     }
-    res.status(404).json({ message: "Enter all the required fields" });
-    return;
+    res.status(404).json({ message: 'Enter all the required fields' })
+    return
   }
-  res.status(404).send("You do not have acess to save the hall");
+  res.status(404).send('You do not have acess to save the hall')
 }
 
 // //Helps the admin to edit the details of the halls and save it again in the database
 
 async function editHall(req, res) {
-  const { id } = req.params;
-  const hall = await HallModel.findOneAndUpdate({ hall_id: id }, req.body);
+  const { id } = req.params
+  const hall = await HallModel.findOneAndUpdate({ hall_id: id }, req.body)
   if (!hall) {
-    res.status(200).send({ message: "Error! Invalid Hall ID" });
+    res.status(200).send({ message: 'Error! Invalid Hall ID' })
   } else {
-    const updateHall = await HallModel.findOne({ hall_id: id });
-    res.status(200).json(updateHall);
+    const updateHall = await HallModel.findOne({ hall_id: id })
+    res.status(200).json(updateHall)
   }
 }
 
 async function getHalls(req, res) {
-  const halls = await HallModel.find({});
+  const halls = await HallModel.find({})
   if (halls) {
-    res.status(200).json(halls);
-    return;
+    res.status(200).json(halls)
+    return
   }
 }
 
 async function getHallsById(req, res) {
-  const { id } = req.params;
-  console.log(id);
-  const hall = await HallModel.findOne({ hall_id: id });
+  const { id } = req.params
+  console.log(id)
+  const hall = await HallModel.findOne({ hall_id: id })
   if (!hall) {
-    res.status(404).json({ message: "Error! Invalid Hall ID" });
+    res.status(404).json({ message: 'Error! Invalid Hall ID' })
   } else {
-    res.status(200).json(hall);
+    res.status(200).json(hall)
   }
 }
 
 async function deleteHall(req, res) {
-  const { hall_id } = req.params;
-  console.log(hall_id);
-  const hall = await HallModel.findOneAndDelete(hall_id);
+  let err = await verify1(req)
+  if (err == 1) {
+    res.status(404).send({ message: 'Token mismatch' })
+    return
+  }
+  const { hall_id } = req.params
+  console.log(hall_id)
+  const hall = await HallModel.findOneAndDelete(hall_id)
   if (!hall) {
-    res.status(200).json({ message: "Error! Invalid Hall ID" });
+    res.status(200).json({ message: 'Error! Invalid Hall ID' })
   } else {
-    res.status(200).json({ message: "Hall Deleted With Given Id" });
+    res.status(200).json({ message: 'Hall Deleted With Given Id' })
   }
 }
 
@@ -110,4 +117,4 @@ module.exports = {
   editHall,
   deleteHall,
   getHallsById,
-};
+}
